+++ packages/state-manager/src/utils/tests/factories.ts
@@ -49,6 +49,9 @@ import {
   HCaptchaFormResponse,
   HCaptchaRequest,
   InviteResultWithSalt,
+  FileMessage,
+  FileEncryptionMetadata,
+  UserProfilesUpdatedPayload,
 } from '@quiet/types'
 import { InviteResult } from '@localfirst/auth'
 import { createLogger } from '../logger'
@@ -66,6 +69,7 @@ import { errorsActions } from '../../sagas/errors/errors.slice'
 import { errorsSelectors } from '../../sagas/errors/errors.selectors'
 import { connectionActions } from '../../sagas/appConnection/connection.slice'
 import { connectionSelectors } from '../../sagas/appConnection/connection.selectors'
+import { randomBytes } from 'crypto'
 
 const logger = createLogger('factories')
 
@@ -125,11 +129,38 @@ export const getBaseTypesFactory = async () => {
     bio: factory.sequence('UserProfileDisplayData.bio', (n: number) => `bio_${n}`),
   })
 
+  factory.define<FileMessage>('FileMessage', Object, {
+    id: factory.sequence('FileMessage.id', (n: number) => `profile-photo-user-profile-photo-cid-${n}-${n}`),
+    channelId: '__profile-photo__',
+  })
+
+  factory.define<FileEncryptionMetadata>('FileEncryptionMetadata', Object, {
+    header: factory.sequence('FileEncryptionMetadata.header', (n: number) => randomBytes(32).toString('base64')),
+    recipient: {
+      generation: 0,
+      type: 'ROLE',
+      name: 'MEMBER',
+    },
+  })
+
+  factory.define<FileMetadata>('FileMetadata', Object, {
+    cid: factory.sequence('FileMetadata.cid', (n: number) => `user-profile-photo-cid-${n}`),
+    path: factory.sequence('FileMetadata.path', (n: number) => `/foo/bar/user-profile-photo-cid-${n}.png`),
+    ext: '.png',
+    name: factory.sequence('FileMetadata.name', (n: number) => `user-profile-photo-name-${n}`),
+    message: factory.assoc('FileMessage'),
+    size: factory.sequence('FileMetadata.size', (n: number) => 1024 + n),
+    width: factory.sequence('FileMetadata.width', (n: number) => 100 + n),
+    height: factory.sequence('FileMetadata.height', (n: number) => 100 + n),
+    enc: factory.assoc('FileEncryptionMetadata'),
+  })
+
   factory.define<UserProfile>('UserProfile', Object, {
     userId: factory.sequence('UserProfile.userId', (n: number) => `userId_${n}`),
     nickname: factory.sequence('UserProfile.nickname', (n: number) => `userProfile.nickname_${n}`),
-    photo: 'dGVzdAo=',
+    photo: undefined,
     bio: factory.sequence('UserProfile.bio', (n: number) => `bio_${n}`),
+    profilePhoto: factory.assoc('FileMetadata'),
   })
 
   factory.define<User>('User', Object, {
@@ -588,6 +619,22 @@ export const getSocketFactory = async () => {
         peerId: 'peer-id',
       },
     },
+  })
+
+  factory.define<UserProfilesUpdatedPayload>(SocketActions.USER_PROFILES_UPDATED, Object, {
+    new: [
+      {
+        userId: 'user-id',
+        nickname: 'Test User',
+        photo: 'dGVzdAo=',
+        bio: 'This is a test user profile',
+        userData: {
+          onionAddress: 'test.onion',
+          peerId: 'peer-id',
+        },
+      },
+    ],
+    updates: [],
   })
 
   factory.define<SetUserProfileResponse>(`${SocketActions.SET_USER_PROFILE}_response`, Object, {